new1.3

+++ FVCalc.java
@@ -13,9 +13,8 @@ public class FVCalc {
 
     // prints
 
-    System.out.println(" After " + years + " years, $" + currentSum + " saved at " + interestRate + "% will yield $" + futureValue);
+    System.out.println(" After " + years + " years, $" + currentSum + " saved at " + interestRate + "% will yield $" + (int)futureValue);
     
 
     }
-    
 }
